+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Bug fixes
 
 - Fixed an error with the keyboard bindings. ([ticket #89](https://wow.curseforge.com/projects/total-rp-3/issues/89))
+- Fixed an error in the Prat module with non-customized names. ([ticket #95](https://wow.curseforge.com/projects/total-rp-3/issues/95))
 
 ## Other improvements